refactor: int type

+++ src/sysconf.c
@@ -1,8 +1,8 @@
 #include "sysconf.h"
 
 plotcfg_t pcfg;
-rrtinfo_t info_r = (rrtinfo_t){.histi = ~(size_t)0};
-rtinfo_t info_c = (rtinfo_t){.histi = ~(size_t)0};
+rrtinfo_t info_r = (rrtinfo_t){.histi = ~0UL};
+rtinfo_t info_c = (rtinfo_t){.histi = ~0UL};
 
 plotcfg_t get_plotcfg() { return pcfg; }
